+++ compile.js
@@ -1,21 +1,25 @@
-/*const fs = require('fs');
-const path = require('path');
-const pug = require('pug');
-const open = require('open');
+import { compileFile } from 'pug';
+import { readdir, writeFile } from 'node:fs/promises';
 
-const indexJsPath = path.join(__dirname, 'src/index.js');
+const PATHS = ['./src/js/views/'];
+  
+const PATH_OUT = './src/Compile/';
 
-const pugDir = path.join(__dirname, 'src');
+for (const path of PATHS) {
+    console.log("in for");
+  readdir(path)
+    .then(async (res) => {
+      for (const fileName of res) {
+        console.log("in for");
+        console.log(fileName);
+        if (fileName.split('.')[1] === 'pug') {
+            console.log("in if");
+          const name = fileName.replace('.pug', '');
 
-const indexHtmlPath = path.join(__dirname, 'src/index.html');
+          const templateFunction = compileFile(path + fileName);
 
-const { App, routes } = require(indexJsPath);
-
-const template = pug.compileFile(path.join(pugDir, 'index.pug'));
-
-const app = new App(routes);
-const html = template({ app });
-
-fs.writeFileSync(indexHtmlPath, html);
-
-open(indexHtmlPath);*/
+          writeFile(PATH_OUT + name + '.js', 'export default ' + templateFunction);
+        }
+      }
+    });
+}
