+++ src/Compile/Feed_v2.js
@@ -1,0 +1,10 @@
+export default function template(locals) {var pug_html = "", pug_mixins = {}, pug_interp;var pug_debug_filename, pug_debug_line;try {;pug_debug_line = 1;pug_debug_filename = ".\u002Fsrc\u002Fjs\u002Fviews\u002FFeed_v2.pug";
+pug_html = pug_html + "\u003Cdiv class=\"feed\"\u003E";
+;pug_debug_line = 2;pug_debug_filename = ".\u002Fsrc\u002Fjs\u002Fviews\u002FFeed_v2.pug";
+pug_html = pug_html + "\u003Ch1\u003E";
+;pug_debug_line = 2;pug_debug_filename = ".\u002Fsrc\u002Fjs\u002Fviews\u002FFeed_v2.pug";
+pug_html = pug_html + "Feed Page\u003C\u002Fh1\u003E";
+;pug_debug_line = 3;pug_debug_filename = ".\u002Fsrc\u002Fjs\u002Fviews\u002FFeed_v2.pug";
+pug_html = pug_html + "\u003Cp\u003E";
+;pug_debug_line = 3;pug_debug_filename = ".\u002Fsrc\u002Fjs\u002Fviews\u002FFeed_v2.pug";
+pug_html = pug_html + "This is the feed page content.\u003C\u002Fp\u003E\u003C\u002Fdiv\u003E";} catch (err) {pug.rethrow(err, pug_debug_filename, pug_debug_line);};return pug_html;}
