+++ src/app/App.js
@@ -1,18 +1,71 @@
+import templateFeed from '../Compile/Feed_v2.js';
+
+function renderFeed (parent) {
+	parent.innerHTML = '';
+	parent.innerHTML = templateFeed();
+};
+
+function renderLogin (parent) {
+	parent.innerHTML = '';
+	parent.innerHTML = `
+            <div class="container">
+                <div class="form-section">
+                    <h1 class="description"><Логин></h1>
+                    <p>Здесь будет информация о пользователе.</p>
+                    <div class="card-actions">
+                    <button class="btn custom-btn" id="dislike">
+                        <img src="../assets/img/X.svg" alt="X">
+                    </button>
+                    <button class="btn custom-btn" id="like">
+                        <img src="../assets/img/Heart.svg" alt="Heart">
+                    </button>
+                    </div>
+                </div>
+                <div class="image-section">
+                <img src="../assets/img/image.svg" alt="Image">
+                </div>
+            </div>
+        `;
+};
+
+function renderSignup (parent) {
+	parent.innerHTML = '';
+	parent.innerHTML = `
+            <div class="container">
+                <div class="form-section">
+                    <h1 class="description"><Логин></h1>
+                    <p>Здесь будет информация о пользователе.</p>
+                    <div class="card-actions">
+                    <button class="btn custom-btn" id="dislike">
+                        <img src="../assets/img/X.svg" alt="X">
+                    </button>
+                    <button class="btn custom-btn" id="like">
+                        <img src="../assets/img/Heart.svg" alt="Heart">
+                    </button>
+                    </div>
+                </div>
+                <div class="image-section">
+                <img src="../assets/img/image.svg" alt="Image">
+                </div>
+            </div>
+        `;
+};
+
 export const createRouter = (root) => ({
     feed: {
       path: '/feed',
       handle: 'Главная',
-      componentName: new MainPage(root),
+      componentName: renderFeed,
     },
     login: {
       path: '/login',
-      handle: 'Авторизоваться',
-      componentName: new LoginPage(root),
+      handle: 'Авторизация',
+      componentName: renderLogin,
     },
     signup: {
       path: '/signup',
       handle: 'Регистрация',
-      componentName: new RegistrationPage(root)
+      componentName: renderSignup
     },
 });
 
@@ -20,10 +73,8 @@ export default class App {
 	#state = {};
 	handlers = {};
 	#structure = {};
-	config;
 	root;
-	constructor(config, root) {
-		this.config = config;
+	constructor(root) {
 		this.root = root;
 		this.#state.user = null;
         this.router = createRouter(root);
@@ -34,11 +85,12 @@ export default class App {
         if (route) {
             history.pushState({}, '', route.path);
             this.clear(true);
-            route.componentName.render();
+            route.componentName(root);
             this.#structure[route.handle] = route.componentName;
         } else {
             history.pushState({}, '', this.router.feed.path);
-            this.#renderFeed();
+			this.goToPage(this.router.login.path);
+           // this.#renderLogin();
         }
 	}
 
@@ -56,7 +108,7 @@ export default class App {
 		});
 	}
 
-	#renderMenu() {
+	/*#renderMenu() {
 		const menu = new Menu(this.config.homeConfig.menu, this.root);
 		if (!this.#structure.menu) {
 			this.#structure.menu = menu;
@@ -194,5 +246,5 @@ export default class App {
 			this.goToPage(router.signup, true);
 		};
 		toSignupLink.addHandler('click', clickHandler);
-	}
+	}*/
 }
