+++ src/components/Navbar/Navbar.js
@@ -1,0 +1,16 @@
+export class Navbar {
+    constructor(parent) {
+        this.parent = parent;
+        this.parent.innerHtml = '';
+        this.node = this.render();
+        this.parent.appendChild(this.node);
+    }
+  
+    render() {
+        return template();
+    }
+  
+    destroy() {
+        this.node.remove();
+    }
+}
