+++ src/components/SwipeCard/SwipeCard.js
@@ -1,0 +1,16 @@
+export class SwipeCard {
+    constructor(parent) {
+      this.parent = parent;
+      this.parent.innerHtml = '';
+      this.node = this.render();
+      this.parent.appendChild(this.node);
+    }
+  
+    render() {
+      return template();
+    }
+  
+    destroy() {
+      this.node.remove();
+    }
+}
