+++ src/index.html
@@ -7,7 +7,6 @@
     <link rel="stylesheet" href="./css/index.css">
 </head>
 <body>
-    <nav id="nav-bar" class="nav"></nav>
     <div id="root"></div>
     <script type="module" src="index.js"></script>
 </body>
